backend/requirements.txt
fastapi==0.116.1
groq==0.31.0
joblib==1.5.1
llvmlite==0.44.0
openai-whisper==20250625
piper-tts==1.3.0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.4
scikit-learn==1.3.2
pydantic==2.11.7
scipy==1.11.4
sentence-transformers==5.1.0
setuptools==80.9.0
sounddevice==0.5.2
uvicorn==0.35.0
websockets==15.0.1
jinja2
pypdf>=5.0.0
python-docx>=1.1.0